feat : Flavor 클래스들을 flavor 디렉토리로 이동
