+++ BE/src/main/java/bibi/demo/domain/flavor/Flavor.java
@@ -1,5 +1,12 @@
-package bibi.demo.domain;
+package bibi.demo.domain.flavor;
 
+import bibi.demo.domain.Image;
+import bibi.demo.domain.Info;
+import bibi.demo.domain.OnSale;
+import bibi.demo.domain.flavor.FlavorAllergen;
+import bibi.demo.domain.flavor.FlavorBase;
+import bibi.demo.domain.flavor.FlavorSyrup;
+import bibi.demo.domain.flavor.FlavorTopping;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
@@ -28,8 +35,6 @@ public class Flavor {
 
     @JsonProperty("is_signature")
     private boolean isSignature;
-    @JsonProperty("is_discontinued")
-    private boolean isDiscontinued;
 
     @OneToOne
     private Info info;
@@ -56,13 +61,12 @@ public class Flavor {
 //
 //    }
 
-    public Flavor(Long id, String nameKR, String nameEN, int kcal, boolean isSignature, boolean isDiscontinued){
+    public Flavor(Long id, String nameKR, String nameEN, int kcal, boolean isSignature){
         this.id = id;
         this.nameKR = nameKR;
         this.nameEN = nameEN;
         this.kcal = kcal;
         this.isSignature = isSignature;
-        this.isDiscontinued = isDiscontinued;
     }
 }
